pyproject.toml
[tool.poetry]
name = "web-hw-module-13"
version = "0.1.0"
description = ""
authors = ["SerhiiAndreiko <130772110+SerhiiAndreiko@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
pytest = "^7.4.3"
requests = "^2.31.0"
httpx = "^0.25.2"


[tool.poetry.group.dev.dependencies]
aiosmtplib = {version = "2.0.2", allow-prereleases = true}
alembic = {version = "1.12.1", allow-prereleases = true}
annotated-types = {version = "0.6.0", allow-prereleases = true}
anyio = {version = "3.7.1", allow-prereleases = true}
bcrypt = {version = "4.1.1", allow-prereleases = true}
blinker = {version = "1.7.0", allow-prereleases = true}
certifi = {version = "2023.11.17", allow-prereleases = true}
click = {version = "8.1.7", allow-prereleases = true}
cloudinary = {version = "1.36.0", allow-prereleases = true}
dnspython = {version = "2.4.2", allow-prereleases = true}
ecdsa = {version = "0.18.0", allow-prereleases = true}
email-validator = {version = "2.1.0.post1", allow-prereleases = true}
fastapi = {version = "0.104.1", allow-prereleases = true}
fastapi-jwt-auth = {version = "0.5.0", allow-prereleases = true}
fastapi-limiter = {version = "0.1.5", allow-prereleases = true}
fastapi-mail = {version = "1.4.1", allow-prereleases = true}
greenlet = {version = "3.0.1", allow-prereleases = true}
h11 = {version = "0.14.0", allow-prereleases = true}
idna = {version = "3.6", allow-prereleases = true}
jinja2 = {version = "3.1.2", allow-prereleases = true}
libgravatar = {version = "1.0.4", allow-prereleases = true}
mako = {version = "1.3.0", allow-prereleases = true}
markupsafe = {version = "2.1.3", allow-prereleases = true}
passlib = {version = "1.7.4", allow-prereleases = true}
psycopg2-binary = {version = "2.9.9", allow-prereleases = true}
pyasn1 = {version = "0.5.1", allow-prereleases = true}
pydantic = {version = "2.5.2", allow-prereleases = true}
pydantic-settings = {version = "2.1.0", allow-prereleases = true}
pydantic-core = {version = "2.14.5", allow-prereleases = true}
pyjwt = {version = "1.7.1", allow-prereleases = true}
python-dotenv = {version = "1.0.0", allow-prereleases = true}
python-jose = {version = "3.3.0", allow-prereleases = true}
python-multipart = {version = "0.0.6", allow-prereleases = true}
redis = {version = "4.6.0", allow-prereleases = true}
rsa = {version = "4.9", allow-prereleases = true}
six = {version = "1.16.0", allow-prereleases = true}
sniffio = {version = "1.3.0", allow-prereleases = true}
sqlalchemy = {version = "2.0.23", allow-prereleases = true}
starlette = {version = "0.27.0", allow-prereleases = true}
typing-extensions = {version = "4.8.0", allow-prereleases = true}
urllib3 = {version = "2.1.0", allow-prereleases = true}
uvicorn = {version = "0.24.0.post1", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
